improve step generation
This change improves a couple of things

### better group matching.

There was an imperfection in the regex that made groups too eager. Before `I do "x" and "y"` would have been transformed to `I do (.*)`, while in theory should be `I do (.*) and (.*)`

### show actual strings instead of generic group

This is more opinable, and maybe should go behind some option flag, but if I have a test that says

```
When I click the button "Save"
and I click the button "Close"
and I click the button "Expand"
...
```

the corresponding autogenerated feature would be something like


```ts
when('I click the button (.*)', (args0) => {});
when('I click the button (.*)', (args0) => {});
when('I click the button (.*)', (args0) => {});
```

making it a little hard to understand what's going on just by looking at the test. This change instead generates the following:

```ts
when('I click the button "(Save)"', (args0) => {});
when('I click the button "(Close)"', (args0) => {});
when('I click the button "(Expand)"', (args0) => {});
```

This equally provides `args0` with the quoted content, but also provide a better textual represenation in the the step file

Let me know what you think :) After feedback, I'll make sure to update/add tests